requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.23.0
polars>=0.19.0

# Machine Learning & NLP
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# Utilities
joblib>=1.2.0
rapidfuzz>=3.0.0
python-dotenv>=0.20.0

# Text Processing
nltk>=3.8

# Visualization (for notebooks)
matplotlib>=3.5.0
seaborn>=0.12.0
